Edit pie chart Data diganti dengan request ajax form edit client dipanggil dengan $_GET['id_client'] dari dashboard tabel flexgrid.

+++ proses/proses_hapus_client.php
@@ -1,9 +1,12 @@
 <?php 
 	include('../koneksi_db/koneksi.php');
 	$id_client = $_GET['id_client'];
-	$sql = mysqli_query($conn, "DELETE FROM client WHERE id_client = '$id_client'");
-
-	mysqli_close();
+	if (isset($_GET['id_client'])) {
+		$id_client = $_GET['id_client'];
+		$sql = mysqli_query($conn, "DELETE FROM client WHERE id_client = '$id_client'");
+		mysqli_close();
+	}
+	
 	header("Location: ../net_monitoring.php?page=1");
  ?>
